server/requirements.txt
aio-pika>=8.0.3
fastapi>=0.78.0
python-dotenv>=0.20.0
uvicorn>=0.18.2
